adding shiny dashboard code for reference
The shiny dashboard is deployed from a separate repo in my github, but this will provide a reference for the code that was used to create it. Since the code for deploying the app is commented out, it should not interfere with site building. However, if you have any issues, try commenting out the following lines of code (46-47):

selectInput("variable_type", label = h3("Family planning variable"),
            choices = type_variable, selected = "Abortion Rate")

Or, you can set everything to 'eval=FALSE'